Initial Unity Build
This was the first game build to get the basic starter files needed to start implementing the web game into my website.
The build included TemplateData, StreamingAssets, Build files, and a default index.html page with boilerplate code.
The original index.html page will stay in the wwwroot/unity folder of the project, however I saved a copy of the page as a Razor view, which will become my Game/Index view in my MVC project.