Base MVC Setup
Created 5 Models:
-Achievement
-Game
-Score
-User
-UserAchievement
The models properties have also been configured in this commit

Created relevant Controllers. The controllers are still a WIP and do not have CRUD functionality yet.

Setup some base Views
Since I started this project as a completely empty MVC project, I noticed some convenient views (and folders) missing that aren't provided in an empty MVC project (such as the wwwroot folder, _ViewStart and _ViewImports razor pages, as well as a Home Index and controller, and Shared folder and _Layout razor page. Those were all created from scratch.)

The Game\Index razor view is a copy of the Unity build's default index.html.
All I changed here was a small section in the script tag to point the "var buildUrl" in the right path of the unity build files.

The Home\Index razor view is mostly code from the free template. I started personalizing a lot of the elements already. I also took a hacksaw to the html code in order to implement this page as a razor view that utilizes the shared _Layout.cshtml.