Implemented Login functionality
Added a Login action to the UserController which takes in a LoginViewModel
Added a new view model, "LoginViewModel"
Fixed the code formatting in the GameIndexScripts (not sure why, but its readability got jumbled moving it from the Game\Index to its own script file)
Finished implementing the script that handles the sign up form functionality and changed some of the comments
Added new script that handles the login form functionality.

To Do:
-Now that users can create accounts and sign in, I need to add UI that shows that they are logged in ( eg:
    -nav bar "sign in" button becomes "log out" or "sign out"
    -a user badge appears in the upper right that when clicked, has a nav menu to change user info such as email or password
    -attempting to change user info redirects to a new view where users can validate their credentials before changing their info
)
-Implement admin functionality such as managing users and leaderboard scores.
    -logging into an admin account reveals UI options only visible when signed on as an admin

There is more I want to do, but considering my time frame I have to focus on some core functionality of the back end system before anything else.