Settings Page Update
-Profile and Settings were combined into one view. There is also a settings.css stylesheet to style the page and settings.js file for some of the functionality on the page.
-The settings form has been split in two to handle updating either username/email or password. Splitting the form up helped simplify the code because it was getting overly complex trying to handle so many actions in one "save changes" button.
-Added a button bar below the settings/profile container. User's can now navigate back to the game index with ease, view the scores they've achieved, or choose to delete their account.
-Updated the user model with a new property "public bool IsDeleted { get; set; }". This will be used to help implement user deletion. Added a migration as a result of the update, then updated the database.
-Updated the ScoreController! The score controller now has a basic structure to it and a working Index() action. This action is used to populate a table in the settings view so users can view their game scores.

WIP:
The delete function is still being developed, there are quite a lot of quirks to figure out with simulated/soft deletion.
Some buttons in the delete confirmation modal still don't work (the cancel and close button).

To Do:
Add another button in the settings button bar to view achievements. This would essentially do the same thing the view scores button does but for achievements.
Implement the profile cards "Highest Score" and "Achievements Earned" quick stats.
Implement Admin... everything. Admin UI that is visible when an admin logs in. Maybe an Admin view that acts as a front end user manager that can optionally manage other data as well such as scores.
And last... develop the web game to an actually playable state.